bank-chatbot/backend-local-llm/requirements.txt
flask
pyodbc
sentence-transformers
faiss-cpu 
faiss-gpu
pymupdf
numpy
transformers
torch